Create students review
Following "This site is open source. Improve this page." comment on the bottom of the website, I suggest to open a section inside the LINKS at "about-me.md" file with student reviews. This way students learn git at the same time that they leave a review during their time learning Economics for business.

And now, back to study!